+++ src/discovery/zeroconf-register.py
@@ -1,0 +1,37 @@
+#!/usr/bin/env python3
+"""
+Example of announcing a service (in this case, a fake HTTP server)
+https://github.com/jstasiak/python-zeroconf/blob/master/examples/registration.py
+"""
+import sys
+import socket
+import logging
+from time import sleep
+from zeroconf import ServiceInfo, Zeroconf
+
+
+if __name__ == '__main__':
+    logging.basicConfig(level=logging.DEBUG)
+    if len(sys.argv) > 1:
+        assert sys.argv[1:] == ['--debug']
+        logging.getLogger('zeroconf').setLevel(logging.DEBUG)
+
+    desc = {'path': '/~tunfish/'}
+
+    info = ServiceInfo("_http._tcp.local.",
+                       "Tunfish's Test Web Site._http._tcp.local.",
+                       socket.inet_aton("127.0.0.1"), 80, 0, 0,
+                       desc, "ash-2.local.")
+
+    zeroconf = Zeroconf()
+    print("Registration of a service, press Ctrl-C to exit...")
+    zeroconf.register_service(info)
+    try:
+        while True:
+            sleep(0.1)
+    except KeyboardInterrupt:
+        pass
+    finally:
+        print("Unregistering...")
+        zeroconf.unregister_service(info)
+        zeroconf.close()